requirements.txt
altgraph==0.17.3
numpy==1.23.5
Pillow==9.3.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
